Added support for "table" and "container" directives.
Added support for the "table" directive.  Insert title into output
with new rststyle-table-title style.

Added limited support for "container" directive.  Limitations: (1)
Only the first class in the list of classes (arguments) is used
and (2) that class/style must be a paragraph style and not (for
example) a character style.

Subversion-Revision: r4852