HTML5: Semantic markup for footnotes.
Use "aside" instead of a description list for footnotes and citations.
While providing decent standard rendering, footnote label and text
are no instance of a "definition term" and "definition".

Use "invisible" footnote label brackets instead of display="none"
with superscript footnotes so that the brackets are kept with
drag-and-drop.

Adapt stylesheets.

Subversion-Revision: r8773