Import ``DocutilsTestSupport`` from ``test``
This is the first step of moving to the standard ``unittest`` style.
It involves making the ``test/`` directory a psuedo-package and
importing the ``DocutilsTestSupport`` module from it, instead of the
path manipulation previously carried out in ``__init__.py`` files.
As of this, we also blank out all ``__init__.py`` files in the test
directories.

Subversion-Revision: r9140