Use absolute paths in ``test_CLI``
When running tests in working directories other than ``docutils/test``,
the tests fail as the expected files cannot be opened from the
relative ``./data`` path. This change uses the defined ``DATA_ROOT``
constant so that the tests pass regardless of working directory. We
also take the opportunity to explicitly specify the encoding with
which to open the file, avoiding platform-dependent behaviour.

Subversion-Revision: r9330