latex2mathml: better error reports.
Keep a copy of the math input for error reporting.

Don't merge runs of whitespace in the math input to a single space:
enables reporting error spot.

Raise MathError error when converting TeX to MathML leaves nodes with less
than required child-elements.
(This is usually a sign of a TeX command missing an argument.)

Subversion-Revision: r9528