Sort "nodes" and "test_nodes".
Sort definitions in "nodes.py" to get a clearer structure.

Move `TextElement` to the end of "Element Categories", so that
the Inline Element Category class can be declared as valid child
element (for use with the to-be-defined `Element.validate()` method.

Sort tests to better match the structure of "nodes.py".

Subversion-Revision: r9680