Doctree validation: custom exception `nodes.ValidationError`.
The custom `ValidationError` has the additional attribute "problematic
element" that holds the element close to the violation ("self", for
invalid XML attributes and spurious text or a child node that does not
fit in the content model).

Use for warnings with source-code line
in `transforms.universal.Validate`

Revise validation error messages.

Subversion-Revision: r9723