xml-parser: Unindent continuation lines. Fixes xml round trip.
The XML writer supports formatting XML with newlines and indentation
(see the corresponding configuration settings).

In most reStructuredText markup blocks, indentation has a semantic
importance and is removed before passing the block for further parsing.
As a result, only a limited set of elements may have text with
spaces after a newline.

Remove indentation of continuation lines unless the containing element
is a `FixedTextElement`, `literal`, or `Unknown` instance.

Subversion-Revision: r9743